Refactor collection managers and catalog utilities
