+++ src/app/_components/collection-links-manager/types.ts
@@ -1,0 +1,9 @@
+export type CollectionLinkModel = {
+  id: string;
+  name: string;
+  url: string;
+  comment: string | null;
+  order: number;
+};
+
+export type Feedback = { type: "success" | "error"; message: string } | null;
